added preshuffle v2
