+++ example/ck_tile/03_gemm/gemm_utils.hpp
@@ -249,7 +249,7 @@ struct GemmConfigPreshufle_1 : public GemmConfigBase
 
     static constexpr int kBlockPerCu           = 2;
     static constexpr auto Scheduler            = ck_tile::GemmPipelineScheduler::Default;
-    static constexpr ck_tile::index_t Pipeline = CK_TILE_PIPELINE_PRESHUFFLE_V3;
+    static constexpr ck_tile::index_t Pipeline = CK_TILE_PIPELINE_PRESHUFFLE_V2;
     static constexpr bool Preshuffle           = true;
     static constexpr bool DoubleSmemBuffer     = true;
 };
@@ -271,7 +271,7 @@ struct GemmConfigPreshufle_2 : public GemmConfigBase
 
     static constexpr int kBlockPerCu           = 2;
     static constexpr auto Scheduler            = ck_tile::GemmPipelineScheduler::Default;
-    static constexpr ck_tile::index_t Pipeline = CK_TILE_PIPELINE_PRESHUFFLE_V3;
+    static constexpr ck_tile::index_t Pipeline = CK_TILE_PIPELINE_PRESHUFFLE_V2;
     static constexpr bool Preshuffle           = true;
     static constexpr bool DoubleSmemBuffer     = true;
 };
@@ -449,15 +449,15 @@ struct PipelineTypeTraits<CK_TILE_PIPELINE_PRESHUFFLE_V2>
         ck_tile::BaseWeightPreshufflePipelineAGmemBGmemCRegV2<PipelineProblem>;
 };
 
-template <>
-struct PipelineTypeTraits<CK_TILE_PIPELINE_PRESHUFFLE_V3>
-{
-    template <typename PipelineProblem>
-    using GemmPipeline = ck_tile::WeightPreshufflePipelineAGmemBGmemCRegV3<PipelineProblem>;
-    template <typename PipelineProblem>
-    using UniversalGemmPipeline =
-        ck_tile::BaseWeightPreshufflePipelineAGmemBGmemCRegV3<PipelineProblem>;
-};
+// template <>
+// struct PipelineTypeTraits<CK_TILE_PIPELINE_PRESHUFFLE_V3>
+// {
+//     template <typename PipelineProblem>
+//     using GemmPipeline = ck_tile::WeightPreshufflePipelineAGmemBGmemCRegV3<PipelineProblem>;
+//     template <typename PipelineProblem>
+//     using UniversalGemmPipeline =
+//         ck_tile::BaseWeightPreshufflePipelineAGmemBGmemCRegV3<PipelineProblem>;
+// };
 
 auto create_args(int argc, char* argv[])
 {
